feat: create/update persistent assetdb queries (ENG-5682) (#12)
[ENG-5682](https://stacklet.atlassian.net/browse/ENG-5682)

### what & why

High-level:
- adds `assetdb_query_save` tool, so that queries can be tweaked and
shared.

But also:
- breaks out `platform` and `assetdb` packages, for organization and
locality of reference.
- both their tools are now registered explicitly in the `mcp`
constructor
- leaves the `docs` tool implementations untouched, in the hope of
avoiding further conflicts in #11
  - assetdb has a bit of additional pydantic typing
- lazy auth/client loading instead of eager middleware, for tastefulness
and locality of reference.
- will conflict with #11 but hopefully not too painfully, just do
`get(ctx)` like the other clients
- `json_guard` decorator for tool implementations to work around the
situation described in
[jlowin/fastmcp#932](https://github.com/jlowin/fastmcp/issues/932) which
may be _closed_ but is emphatically not _fixed_.

### testing

Lots of unit tests, although plenty of coverage still lagging; got
claude to stress-test it to its own "satisfaction" a few times, which
did turn up a bug or two, which was nice.

### docs

Contains at least some, including CLAUDE updates.


[ENG-5682]:
https://stacklet.atlassian.net/browse/ENG-5682?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ