feat: config/instructions (ENG-5749) (#29)
[ENG-5749](https://stacklet.atlassian.net/browse/ENG-5749)

### what

- More explicit sample configuration files (plus a minor hack to keep
claude zero-config)
- Slightly better documentation of agent integrations
- Dire warnings of the potential implications of using the server
- Couple of bits of nuance I thought the gql info tool needed
- Remove not-very-helpful examples from some tool docs
- Reorder _dataset_info ahead of _export/_lookup

### why

When I wrote ENG-5749, such as it is, the motivation was "configuration
seems a bit weird and opaque" (what with the hidden file…), and I think
this makes it a bit less opaque. And I added the other docs because I
thought of them…

### testing

`just install` writes `.mcp.json` when it's not there, and ignores it
when it is.

### docs

Mostly docs.


[ENG-5749]:
https://stacklet.atlassian.net/browse/ENG-5749?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ